Migrated to platform independent types that are lossless/compatible irrespective of the platform. E.g. SIZE_T, & int32 not int

+++ Source/UnrealSodium/Private/UnrealSodiumBlueprintInterface.cpp
@@ -202,11 +202,11 @@ void UUnrealSodiumPluginBPLibrary::DecryptStringSymmetric(TArray<uint8> encrypte
 	}
 }
 
-int UUnrealSodiumPluginBPLibrary::AsymmetricNonceLength() {
+int64 UUnrealSodiumPluginBPLibrary::AsymmetricNonceLength() {
 	return FUnrealSodiumModule::Get().GetAssymetricNonceLen();
 }
 
-int UUnrealSodiumPluginBPLibrary::SymmetricNonceLength() {
+int64 UUnrealSodiumPluginBPLibrary::SymmetricNonceLength() {
 	return FUnrealSodiumModule::Get().GetSymmetricNonceLen();
 }
