+++ Source/UnrealSodium/Public/UnrealSodium.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "Modules/ModuleManager.h"
+#include "HAL/Platform.h"
+#include "GenericPlatform/GenericPlatform.h"
 #include "../../ThirdParty/libsodium/src/libsodium/include/sodium.h"
 
 class FUnrealSodiumModule : public IModuleInterface
@@ -17,26 +19,26 @@ public:
 	void GenerateKeyPair(TArray<uint8>& publicKey, TArray<uint8>& secretKey);
 	TArray<uint8> GenerateKey();
 
-	int GetPublicKeyBytes() { return crypto_box_PUBLICKEYBYTES; }
-	int GetSecretKeyBytes() { return crypto_box_SECRETKEYBYTES; }
-	int GetBoxSealBytes() { return crypto_box_SEALBYTES; }
-	int GetMacBytes(){ return crypto_box_MACBYTES; }
-	int GetMacBytesSymmetric() { return crypto_secretbox_MACBYTES; }
+	SIZE_T GetPublicKeyBytes() { return crypto_box_PUBLICKEYBYTES; }
+	SIZE_T GetSecretKeyBytes() { return crypto_box_SECRETKEYBYTES; }
+	SIZE_T GetBoxSealBytes() { return crypto_box_SEALBYTES; }
+	SIZE_T GetMacBytes(){ return crypto_box_MACBYTES; }
+	SIZE_T GetMacBytesSymmetric() { return crypto_secretbox_MACBYTES; }
 
-	int GetAssymetricNonceLen() { return crypto_box_NONCEBYTES; }
-	int GetSymmetricNonceLen() { return crypto_secretbox_NONCEBYTES; }
+	SIZE_T GetAssymetricNonceLen() { return crypto_box_NONCEBYTES; }
+	SIZE_T GetSymmetricNonceLen() { return crypto_secretbox_NONCEBYTES; }
 
 	void RandomBytes(unsigned char* bytes, size_t len);
 
-	int Encrypt(TArray<uint8>& encrypted, TArray<uint8>& data, TArray<uint8>& publicKey);
-	int Decrypt(TArray<uint8>& decrypted, TArray<uint8>& encrypted, TArray<uint8>& publicKey, TArray<uint8>& privateKey);
+	int32 Encrypt(TArray<uint8>& encrypted, TArray<uint8>& data, TArray<uint8>& publicKey);
+	int32 Decrypt(TArray<uint8>& decrypted, TArray<uint8>& encrypted, TArray<uint8>& publicKey, TArray<uint8>& privateKey);
 
-	int EncryptAuthenticated(TArray<uint8>& encrypted, TArray<uint8>& data, TArray<uint8>& nonce, TArray<uint8>& publicKey, TArray<uint8>& privateKey);
-	int DecryptAuthenticated(TArray<uint8>& decrypted, TArray<uint8>& encrypted, TArray<uint8>& nonce, TArray<uint8>& publicKey, TArray<uint8>& privateKey);
+	int32 EncryptAuthenticated(TArray<uint8>& encrypted, TArray<uint8>& data, TArray<uint8>& nonce, TArray<uint8>& publicKey, TArray<uint8>& privateKey);
+	int32 DecryptAuthenticated(TArray<uint8>& decrypted, TArray<uint8>& encrypted, TArray<uint8>& nonce, TArray<uint8>& publicKey, TArray<uint8>& privateKey);
 
 
-	int EncryptSymmetric(TArray<uint8>& encrypted, TArray<uint8>& data, TArray<uint8>& nonce, TArray<uint8>& key);
-	int DecryptSymmetric(TArray<uint8>& decrypted, TArray<uint8>& encrypted, TArray<uint8>& nonce, TArray<uint8>& key);
+	int32 EncryptSymmetric(TArray<uint8>& encrypted, TArray<uint8>& data, TArray<uint8>& nonce, TArray<uint8>& key);
+	int32 DecryptSymmetric(TArray<uint8>& decrypted, TArray<uint8>& encrypted, TArray<uint8>& nonce, TArray<uint8>& key);
 
 	static inline FUnrealSodiumModule& Get() {
 		return FModuleManager::LoadModuleChecked<FUnrealSodiumModule>("UnrealSodium"); // name should be the same as directory of the plugin in /Plugins
